% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{add.anova.pairwise.comparison}
\alias{add.anova.pairwise.comparison}
\title{add.anova.pairwise.comparison}
\usage{
add.anova.pairwise.comparison(
  df,
  ymin,
  space,
  spaceLabel,
  lenTip,
  includeComps = c()
)
}
\description{
add.anova.pairwise.comparison
}
