% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{CholeskyFactorCovType}
\alias{CholeskyFactorCovType}
\title{export}
\description{
export

export
}
\section{Super classes}{
\code{\link[swr:BareType]{swr::BareType}} -> \code{\link[swr:CoreType]{swr::CoreType}} -> \code{\link[swr:MatrixType]{swr::MatrixType}} -> \code{CholeskyFactorCovType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CholeskyFactorCovType$new()}}
\item \href{#method-clone}{\code{CholeskyFactorCovType$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="swr" data-topic="CoreType" data-id="set_value">}\href{../../swr/html/CoreType.html#method-set_value}{\code{swr::CoreType$set_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="swr" data-topic="CoreType" data-id="textify">}\href{../../swr/html/CoreType.html#method-textify}{\code{swr::CoreType$textify()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CholeskyFactorCovType$new(
  name,
  transforms = swr::transforms(),
  size = swr::dimensions(0, 0),
  dimensions = swr::dimensions(),
  value = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CholeskyFactorCovType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
