% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-env.R
\name{env_copy}
\alias{env_copy}
\title{Copy objects from one env to another}
\usage{
env_copy(from, to, nms)
}
\arguments{
\item{from}{environment to copy from}

\item{to}{environment to copy to}

\item{nms}{names of objects to copy}
}
\value{
names of objects copied
}
\description{
Ignores missing objects.
}
