% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{InlineBlockType}
\alias{InlineBlockType}
\title{An inline block of Stan code}
\description{
An inline block \emph{can} be merged with another
}
\section{Super class}{
\code{\link[swr:BareBlockType]{swr::BareBlockType}} -> \code{InlineBlockType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineBlockType$new()}}
\item \href{#method-textify_declarations}{\code{InlineBlockType$textify_declarations()}}
\item \href{#method-textify_statements}{\code{InlineBlockType$textify_statements()}}
\item \href{#method-textify}{\code{InlineBlockType$textify()}}
\item \href{#method-clone}{\code{InlineBlockType$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBlockType$new(..., declarations = rlang::env())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textify_declarations"></a>}}
\if{latex}{\out{\hypertarget{method-textify_declarations}{}}}
\subsection{Method \code{textify_declarations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBlockType$textify_declarations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textify_statements"></a>}}
\if{latex}{\out{\hypertarget{method-textify_statements}{}}}
\subsection{Method \code{textify_statements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBlockType$textify_statements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textify"></a>}}
\if{latex}{\out{\hypertarget{method-textify}{}}}
\subsection{Method \code{textify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBlockType$textify()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBlockType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
