% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{is_not_block}
\alias{is_not_block}
\title{Check that 'x' is NOT a Stan block}
\usage{
is_not_block(x)
}
\arguments{
\item{x}{object to check}
}
\value{
TRUE iff 'x' is NOT a BlockType
}
\description{
Check that 'x' is NOT a Stan block
}
