% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{is_statement}
\alias{is_statement}
\title{Check that 'x' is a Stan statement}
\usage{
is_statement(x)
}
\arguments{
\item{x}{object to check}
}
\value{
TRUE iff 'x' is a StatementType
}
\description{
Check that 'x' is a Stan statement
}
