% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{MultiStatementType}
\alias{MultiStatementType}
\title{Class representing a series of Stan statements, not user-facing}
\value{
class representing the statement
}
\description{
Class representing a series of Stan statements, not user-facing

Class representing a series of Stan statements, not user-facing
}
\section{Super class}{
\code{\link[swr:BareStatementType]{swr::BareStatementType}} -> \code{MultiStatementType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiStatementType$new()}}
\item \href{#method-textify_declarations}{\code{MultiStatementType$textify_declarations()}}
\item \href{#method-textify_statements}{\code{MultiStatementType$textify_statements()}}
\item \href{#method-statements}{\code{MultiStatementType$statements()}}
\item \href{#method-declarations}{\code{MultiStatementType$declarations()}}
\item \href{#method-append}{\code{MultiStatementType$append()}}
\item \href{#method-clone}{\code{MultiStatementType$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$new(..., declarations = rlang::env(), indent = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{expressions to treat as Stan statements}

\item{\code{declarations}}{environment to use to create declarations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textify_declarations"></a>}}
\if{latex}{\out{\hypertarget{method-textify_declarations}{}}}
\subsection{Method \code{textify_declarations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$textify_declarations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textify_statements"></a>}}
\if{latex}{\out{\hypertarget{method-textify_statements}{}}}
\subsection{Method \code{textify_statements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$textify_statements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-statements"></a>}}
\if{latex}{\out{\hypertarget{method-statements}{}}}
\subsection{Method \code{statements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$statements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-declarations"></a>}}
\if{latex}{\out{\hypertarget{method-declarations}{}}}
\subsection{Method \code{declarations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$declarations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\if{latex}{\out{\hypertarget{method-append}{}}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$append(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStatementType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
