% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{pick_symbols}
\alias{pick_symbols}
\title{Extract all symbols from an expression as a tree strings}
\usage{
pick_symbols(x)
}
\arguments{
\item{x}{expression to extract symbols from}
}
\value{
a tree (nested lists) of character vectors
}
\description{
The leading guard ("~") is extracted, parenthesis order is
retained but parentheses are not.
}
