% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{pick_variables}
\alias{pick_variables}
\title{Extract all variables used in the code that exist in the environment}
\usage{
pick_variables(code, env)
}
\arguments{
\item{code}{list of expressions to extract variables from}

\item{env}{environment to check for variables}
}
\value{
a named list of the variables and their values as defined
}
\description{
Uses \code{pick_symbols(code)} followed by filtering, undefined symbols are not
returned.
}
