% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{stringify_declarations}
\alias{stringify_declarations}
\title{Transform list of expressions into a list of Stan variable declarations}
\usage{
stringify_declarations(code, env)
}
\arguments{
\item{code}{list of expressions to extract variables from}

\item{env}{environment to check for variables}
}
\value{
a named list of the variables and their corresponding Stan
declarations
}
\description{
Transform list of expressions into a list of Stan variable declarations
}
