% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{stringify_statements}
\alias{stringify_statements}
\title{Transform a list of expressions into a list of Stan code statements}
\usage{
stringify_statements(code)
}
\arguments{
\item{code}{list of expressions to extract variables from}
}
\value{
a named list of the variables and their corresponding Stan
declarations
}
\description{
Transform a list of expressions into a list of Stan code statements
}
