% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{wrap_statements}
\alias{wrap_statements}
\title{Wrap expressions (or statements) into a list of statements}
\usage{
wrap_statements(..., declarations = rlang::env())
}
\arguments{
\item{...}{any expressions or anything descended from BareStatementType}

\item{declarations}{environment containing declared Stan variables}
}
\value{

}
\description{
Works even if the expressions can not be evaluated
}
